requirements.txt
streamlit
google-generativeai
python-dotenv
pdfplumber
python-docx
pandas
numpy
